Consolidate boilerplate linter code
